added some edge case handling

+++ fetcher.js
@@ -1,5 +1,11 @@
 const fs = require('fs');
 const request = require('request');
+const readline = require('readline');
+
+const rl = readline.createInterface({
+  input: process.stdin,
+  output: process.stdout
+});
 
 const args = process.argv.slice(2);
 let adrs = args[0];
@@ -7,21 +13,38 @@ let file = args[1];
 
 
 request(adrs, (error, response, body) => {
-  console.log('error:', error); // Print the error if one occurred
-  console.log('statusCode:', response && response.statusCode); // Print the response status code if a response was received
-  fs.writeFile(file, body, (err) => {
-    if (err) throw err;
-    let fileSize = fs.statSync(file).size;
-    console.log(`Downloaded and saved ${fileSize} bytes to ${file}.`);
-  });
+  if (error) {
+    console.log('error:', error); // Print the error if one occurred
+    
+  }
+  if (fs.existsSync(file)) {
+    rl.question("The file already exists...Overwrite?(y, n)", (answer) => {
+      if (answer === 'y' || answer === 'yes' || answer === 'Yes') {
+        fs.writeFile(file, body, (err) => {
+          if (err) {
+            console.log("invalid file path");
+            process.exit();
+          }
+          let fileSize = fs.statSync(file).size;
+          console.log(`Downloaded and saved ${fileSize} bytes to ${file}.`);
+          process.exit();
+        });
+      } else {
+        console.log("Please choose a new file to write to and rerun... Good bye.");
+        process.exit();
+      }
+
+    });
+  } else {
+    fs.writeFile(file, body, (err) => {
+      if (err) {
+        console.log("invalid file path");
+        process.exit();
+      }
+      let fileSize = fs.statSync(file).size;
+      console.log(`Downloaded and saved ${fileSize} bytes to ${file}.`);
+      process.exit();
+    });
+  }
 });
 
-// const breedDetailsFromFile = function(breed, callback) {
-//   console.log('breedDetailsFromFile: Calling readFile...');
-//   fs.readFile(`./data/${breed}.txt`, 'utf8', (error, data) => {
-//     console.log("In readFile's Callback: it has the data.");
-    
-//     if (!error) callback(data);
-//     else callback(undefined);
-//   });
-// };
